Swift refactoring
